+++ airflow_watchdog/__init__.py
@@ -21,4 +21,4 @@ Install:
     pip install airflow-plugin-watchdog
 """
 
-__version__ = "0.6.2"
+__version__ = "0.6.3"